Initial implementation of build system
Create an initial build system using CMake. For now, only the clients
are built. Also, shared libraries on Windows are not currently supported
due to the lack of export decoration, both locally, and for the LCM
types (the latter needs support in lcmgen, which is not yet implemented
as of this commit).